backend/requirements.txt
fastapi==0.89.1
python-multipart==0.0.5
uvicorn==0.20.0
onnxruntime==1.13.1
numpy==1.24.1
Pillow==9.4.0
python-dotenv==0.21.1
